formula's corrected

+++ include/Stack.h
@@ -5,6 +5,13 @@ enum Operators { SUM, SUB, MUL, DIV, OPEN_BRACKET, CLOSE_BRACKET};
 
 template <typename T> class stakk {
 public:
+	stakk() {}
+
+	void print() {
+		for (int i = 0; i < m_storage.Size(); ++i)
+			cout << m_storage[i] << " ";
+		cout << endl;
+	}
 
 	void push(const T& data) {
 		m_storage.AddData(data);
@@ -19,10 +26,13 @@ public:
 		if (!m_storage.isVoid())
 			return  m_storage.GetValue(m_storage.Size() - 1);
 		else
-			throw std::logic_error("Stack is empty!");
+			throw std::logic_error("1.Stack is empty!");
 	}
 	void pop() {
-		m_storage.deleteElement(m_storage.Size() - 1);
+		if (!empty())
+			m_storage.deleteElement(m_storage.Size() - 1);
+		else
+			throw std::logic_error("2.Stack is empty!");
 	}
 
 private:
